Change type of num argument to long, assert range in function

+++ src/parse.h
@@ -21,17 +21,17 @@ Parse *parse_init(const char *fname);
 void parse_deinit(Parse *file);
 
 // Read one string token into a fixed buffer.
-int parse_string(Parse *file, char *str, int size);
+int parse_string(Parse *file, char *str, long size);
 
 // Read ASCII tokens into typed elements.
-int parse_ascii(Parse *file, void *buf, int num, MPI_Datatype datatype);
+int parse_ascii(Parse *file, void *buf, long num, MPI_Datatype datatype);
 
-int parse_ascii_split(Parse *file, void *buf, int num, MPI_Datatype datatype);
+int parse_ascii_split(Parse *file, void *buf, long num, MPI_Datatype datatype);
 
 // Read binary data into typed elements.
-int parse_binary(Parse *file, void *buf, int num, MPI_Datatype datatype, int swap);
+int parse_binary(Parse *file, void *buf, long num, MPI_Datatype datatype, int swap);
 
-int parse_binary_split(Parse *file, void *buf, int num, MPI_Datatype datatype, int swap);
+int parse_binary_split(Parse *file, void *buf, long num, MPI_Datatype datatype, int swap);
 
 // Dispatch parsing based on the provided mode.
-int parse(Parse *file, void *buf, int num, MPI_Datatype datatype, int mode);
+int parse(Parse *file, void *buf, long num, MPI_Datatype datatype, int mode);
